- Moved Checkpoint handling into Checkpoint Manager
